requirements.txt
rich>=13.7.0
pillow>=10.2.0
python-magic-bin>=0.4.14
